requirements.txt
Flask==3.0.0
reportlab==4.0.7
python-docx==1.1.0
flask-cors==4.0.0
